Address rereview comments
