+++ workflows/prognostic_c48_run/test_prepare_config.py
@@ -112,52 +112,3 @@ def test_get_user_config_is_valid():
     config = prepare_config.user_config_from_dict_and_args(dict_, Args)
     # validate using dacite.from_dict
     dacite.from_dict(UserConfig, dataclasses.asdict(config))
-
-
-def test__physics_frequency_overlay_raises_error():
-    diagnostics = [
-        FortranFileConfig(
-            "physics_3hourly.zarr",
-            {"time": 1},
-            [FortranVariableNameSpec("gfs_phys", "totprcpb_ave", "PRATEsfc")],
-            FortranTimeConfig("interval", 10800),
-        ),
-        FortranFileConfig(
-            "physics_hourly.zarr",
-            {"time": 1},
-            [FortranVariableNameSpec("gfs_sfc", "t2m", "TMP2m")],
-            FortranTimeConfig("interval", 3600),
-        ),
-    ]
-    with pytest.raises(NotImplementedError):
-        prepare_config._physics_frequency_overlay(diagnostics, 900)
-
-
-@pytest.mark.parametrize(
-    "time_config, expected_frequency_in_hours",
-    [
-        (FortranTimeConfig(kind="every"), 0.25),
-        (FortranTimeConfig(kind="interval", frequency=900), 0.25),
-        (FortranTimeConfig(kind="interval", frequency=3600), 1),
-    ],
-)
-def test__physics_frequency_overlay(time_config, expected_frequency_in_hours):
-    diagnostics = [
-        FortranFileConfig(
-            "physics_3hourly.zarr",
-            {"time": 1},
-            [
-                FortranVariableNameSpec("gfs_phys", "totprcpb_ave", "PRATEsfc"),
-                FortranVariableNameSpec("gfs_sfc", "t2m", "TMP2m"),
-            ],
-            time_config,
-        ),
-    ]
-    overlay = prepare_config._physics_frequency_overlay(diagnostics, 900)
-    expected_overlay = {
-        "namelist": {
-            "atmos_model_nml": {"fhout": expected_frequency_in_hours},
-            "gfs_physics_nml": {"fhzero": expected_frequency_in_hours},
-        }
-    }
-    assert overlay == expected_overlay
